requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==2.2.2
    # via
    #   keras
    #   tensorboard
    #   tensorflow
astunparse==1.6.3
    # via tensorflow
beautifulsoup4==4.13.4
    # via gdown
blinker==1.9.0
    # via flask
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via flask
colorama==0.4.6
    # via
    #   click
    #   tqdm
deepface==0.0.93
    # via -r requirements.in
filelock==3.18.0
    # via gdown
fire==0.7.0
    # via deepface
flask==3.1.1
    # via
    #   -r requirements.in
    #   deepface
    #   flask-cors
flask-cors==6.0.0
    # via deepface
flatbuffers==25.2.10
    # via tensorflow
gast==0.6.0
    # via tensorflow
gdown==5.2.0
    # via
    #   deepface
    #   retina-face
google-pasta==0.2.0
    # via tensorflow
grpcio==1.71.0
    # via
    #   tensorboard
    #   tensorflow
gunicorn==23.0.0
    # via deepface
h5py==3.13.0
    # via
    #   keras
    #   tensorflow
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
joblib==1.5.1
    # via mtcnn
keras==3.10.0
    # via
    #   deepface
    #   tensorflow
libclang==18.1.1
    # via tensorflow
lz4==4.4.4
    # via mtcnn
markdown==3.8
    # via tensorboard
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.5.1
    # via
    #   keras
    #   tensorflow
mtcnn==1.0.0
    # via deepface
namex==0.0.9
    # via keras
numpy==2.1.3
    # via
    #   deepface
    #   h5py
    #   keras
    #   ml-dtypes
    #   opencv-python
    #   pandas
    #   retina-face
    #   tensorboard
    #   tensorflow
opencv-python==4.11.0.86
    # via
    #   deepface
    #   retina-face
opt-einsum==3.4.0
    # via tensorflow
optree==0.15.0
    # via keras
packaging==25.0
    # via
    #   gunicorn
    #   keras
    #   tensorboard
    #   tensorflow
pandas==2.2.3
    # via deepface
pillow==11.2.1
    # via
    #   deepface
    #   retina-face
protobuf==5.29.4
    # via
    #   tensorboard
    #   tensorflow
pygments==2.19.1
    # via rich
pysocks==1.7.1
    # via requests
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via -r requirements.in
pytz==2025.2
    # via pandas
requests[socks]==2.32.3
    # via
    #   -r requirements.in
    #   deepface
    #   gdown
    #   tensorflow
retina-face==0.0.17
    # via deepface
rich==14.0.0
    # via keras
six==1.17.0
    # via
    #   astunparse
    #   google-pasta
    #   python-dateutil
    #   tensorboard
    #   tensorflow
soupsieve==2.7
    # via beautifulsoup4
tensorboard==2.19.0
    # via tensorflow
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow==2.19.0
    # via
    #   -r requirements.in
    #   deepface
    #   retina-face
tensorflow-io-gcs-filesystem==0.31.0
    # via tensorflow
termcolor==3.1.0
    # via
    #   fire
    #   tensorflow
tqdm==4.67.1
    # via
    #   deepface
    #   gdown
typing-extensions==4.13.2
    # via
    #   beautifulsoup4
    #   optree
    #   tensorflow
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via requests
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
    #   tensorboard
wheel==0.45.1
    # via astunparse
wrapt==1.17.2
    # via tensorflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools
